test: convert select tests to TypeScript
