+++ tests/token.test.ts
@@ -1,3 +1,4 @@
+import { expect, test } from '@jest/globals';
 import { Token } from '../src/lexer/Token.js';
 
 test('Token.toJSON produces plain object', () => {
